all basic functionality accomplished, integral tests to be done
git-svn-id: https://svn.rice.edu/r/software_0/ObjectQueryModel@3278 522aa1da-077c-40de-a48b-1771163b7009

+++ src/tests/source/TestLA21_Instance.cc
@@ -30,7 +30,7 @@ int main(int argc, char **argv){
 	assert(argc==6);
 	bool printResult = (strcmp(argv[1],"Y")==0);
 	bool clusterMode = (strcmp(argv[2],"Y")==0);
-	int blockSize = atoi(argv[3]);
+	size_t blockSize = atoi(argv[3]);
 	int port = 8108;
 	std::string masterIP = argv[4];
 	FILE * targetCode = fopen(argv[5],"r");
